Added some visuals and sounds. Created a type of block that teleporting isn't allowed to work if hit.
Started creating levels, made a testing room so just use that to chuck things and test on I guess.

If you could add to the script so if it hits a block tagged with "NoTeleport" that would be great. There should be a prefab so just add to that I guess.